+++ src/styles/baseStyles.ts
@@ -1,4 +1,4 @@
-export default {
+export const BASE_STYLES = {
   buttonBorder: '0.1rem solid transparent',
   borderRadius: '1.6rem',
   buttonFontSize: '1rem',
